requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-jazzmin==3.0.0
djangorestframework==3.15.2
drf-yasg==1.21.7
geographiclib==2.0
geopy==2.4.1
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.4
graphql-relay==3.2.0
idna==3.9
inflection==0.5.1
packaging==24.1
phonenumbers==8.13.45
pillow==10.4.0
promise==2.3
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
uritemplate==4.1.1
urllib3==2.2.3
